optimize login process, now login info will expire after 7 days

+++ apigateway/apps/store_coupons_handle.py
@@ -1,12 +1,19 @@
 import requests
 from flask import request, jsonify
+from config import config
 from .base import BaseHandler
 from .json_validate import SCHEMA
+from wxbase.utils import create_md5_key
 
 
 class StoreCouponsHandler(BaseHandler):
 
     def get(self, store_id):
+        flag, tag = self.authenticate(request, store_id,
+                                      create_md5_key(config['secret']))
+        if not flag:
+            return self.error_msg(tag)
+
         params = request.args.to_dict()
         # is_valid, tag = self.validate_dict_with_schema(
         #     params, SCHEMA['store_coupons_get'])
@@ -24,6 +31,11 @@ class StoreCouponsHandler(BaseHandler):
         return jsonify(api_resp.json()), resp_status
 
     def post(self, store_id):
+        flag, tag = self.authenticate(request, store_id,
+                                      create_md5_key(config['secret']))
+        if not flag:
+            return self.error_msg(tag)
+
         is_valid, data = self.get_params_from_request(
             request, SCHEMA['store_coupons_post'])
         if not is_valid:
@@ -43,6 +55,11 @@ class StoreCouponsHandler(BaseHandler):
 class StoreCouponHandler(BaseHandler):
 
     def get(self, store_id, coupon_id):
+        flag, tag = self.authenticate(request, store_id,
+                                      create_md5_key(config['secret']))
+        if not flag:
+            return self.error_msg(tag)
+
         api_resp = requests.get(
             '{0}/transactions/coupons/{1}'.format(
                 self.endpoint['transactions'], coupon_id),
@@ -54,6 +71,11 @@ class StoreCouponHandler(BaseHandler):
         return jsonify(api_resp.json()), resp_status
 
     def put(self, store_id, coupon_id):
+        flag, tag = self.authenticate(request, store_id,
+                                      create_md5_key(config['secret']))
+        if not flag:
+            return self.error_msg(tag)
+
         is_valid, data = self.get_params_from_request(
             request, SCHEMA['store_coupon_put'])
         if not is_valid:
@@ -71,6 +93,11 @@ class StoreCouponHandler(BaseHandler):
         return jsonify(api_resp.json()), resp_status
 
     def delete(self, store_id, coupon_id):
+        flag, tag = self.authenticate(request, store_id,
+                                      create_md5_key(config['secret']))
+        if not flag:
+            return self.error_msg(tag)
+
         api_resp = requests.delete(
             '{0}/transactions/coupons/{1}'.format(
                 self.endpoint['transactions'], coupon_id),
